requirements.txt
scikit-learn
matplotlib
ipykernel
emoji-2.15.0
